+++ pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "MagGeo"
+version = "0.1.0"
+description = "MagGeo is a tool that helps ecologists or animal movement researchers to link earth's magnetic field data from satellite source to GPS trajectories."
+authors = ["mfbenitezp <Fernando.Benitez@st-andrews.ac.uk>"]
+license = "GPL-3.0-only"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+pandas = "^1.4.1"
+viresclient = "0.9.0"
+matplotlib = "^3.5.1"
+chaosmagpy = "0.7.1"
+
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
